implement user authorization

+++ RentSystem.API/Controllers/AdvertController.cs
@@ -1,7 +1,13 @@
 ﻿using FluentValidation;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using RentSystem.API.Extensions;
+using RentSystem.Core.Contracts.Repository;
 using RentSystem.Core.Contracts.Service;
 using RentSystem.Core.DTOs;
+using RentSystem.Core.Enums;
+using RentSystem.Core.Policies;
+using System.Security.Claims;
 
 namespace RentSystem.API.Controllers
 {
@@ -10,10 +16,17 @@ namespace RentSystem.API.Controllers
     public class AdvertController : ControllerBase
     {
         private readonly IAdvertService _advertService;
+        private readonly IAdvertRepository _advertRepository;
+        private readonly IUserService _userService;
+        private readonly IAuthorizationService _authorizationService;
         private readonly IValidator<AdvertDTO> _validator;
-        public AdvertController(IAdvertService advertService, IValidator<AdvertDTO> validator)
+
+        public AdvertController(IAdvertService advertService, IUserService userService, IAdvertRepository advertRepository, IAuthorizationService authorizationService, IValidator<AdvertDTO> validator)
         {
             _advertService = advertService;
+            _advertRepository = advertRepository;
+            _userService = userService;
+            _authorizationService = authorizationService;
             _validator = validator;
         }
 
@@ -30,13 +43,23 @@ namespace RentSystem.API.Controllers
         }
 
         [HttpPost]
+        [AuthorizeRole(Role.Owner)]
         public async Task<IActionResult> Create(AdvertDTO advertDTO)
         {
+            var userId = User.Claims.FirstOrDefault(x => x.Type == "UserId")?.Value;
+
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+
+            var user = await _userService.GetAsync(int.Parse(userId)) ?? throw new Exception("User not found");
+
             var result = _validator.Validate(advertDTO);
 
             if (result.IsValid)
             {
-                await _advertService.CreateAsync(advertDTO);
+                await _advertService.CreateAsync(advertDTO, user.Id);
                 return Ok();
             }
 
@@ -45,9 +68,19 @@ namespace RentSystem.API.Controllers
         }
 
         [HttpPut("{id}")]
+        [AuthorizeRole(Role.Owner)]
         public async Task<IActionResult> Update(int id, AdvertDTO advertDTO)
         {
             var result = _validator.Validate(advertDTO);
+
+            var advert = await _advertRepository.GetAsync(id);
+
+            var authResult = await _authorizationService.AuthorizeAsync(User, advert, PolicyNames.SameUser);
+
+            if (!authResult.Succeeded)
+            {
+                return Forbid();
+            }
 
             if (result.IsValid)
             {
@@ -61,8 +94,18 @@ namespace RentSystem.API.Controllers
         }
 
         [HttpDelete]
+        [AuthorizeRole(Role.Owner)]
         public async Task<IActionResult> Delete(int id)
         {
+            var advert = await _advertRepository.GetAsync(id);
+
+            var authResult = await _authorizationService.AuthorizeAsync(User, advert, PolicyNames.SameUser);
+
+            if (!authResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             await _advertService.DeleteAsync(id);
 
             return Ok();
